requirements.txt
alembic==1.13.2
annotated-types==0.7.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pika==1.3.2
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
SQLAlchemy==2.0.32
starlette==0.38.4
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
